requirements.txt
flask
pandas
sentence-transformers
faiss-cpu
fuzzywuzzy
python-Levenshtein
